% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IntegrateData}
\alias{IntegrateData}
\title{Integrate query cell types  with reference cell types}
\usage{
IntegrateData(ref_path, query_path, Celltype_col)
}
\arguments{
\item{data}{A data frame containing marker columns, a `Celltype` column, and a `batch` column}

\item{markers}{A character vector of column names corresponding to marker features}

\item{ref_batch}{Character value indicating the reference batch label}

\item{query_batch}{Character value indicating the query batch label}
}
\value{
A data frame with quantile-normalized marker values and original `Celltype` column
}
\description{
This function normalizes query data to reference data using PCA-space
quantile normalization, performed separately for each cell type.
}
