% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{Normalize_function}
\alias{Normalize_function}
\title{Normalization function}
\usage{
Normalize_function(target, reference)
}
\arguments{
\item{target}{Matrix to normalize}

\item{reference}{Reference matrix}
}
\value{
normalized matrix
}
\description{
Normalization function
}
\keyword{internal}
