% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{NormalizeData}
\alias{NormalizeData}
\title{PCA-space Normalization}
\usage{
NormalizeData(ct1_rep1_i_sig, file_ref_sig)
}
\arguments{
\item{ct1_rep1_i_sig}{Target matrix}

\item{file_ref_sig}{Reference matrix}
}
\value{
normalized matrix in original (reference) space
}
\description{
PCA-space Normalization
}
\keyword{internal}
