% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellfuse_interface.R
\name{predict_cells_time}
\alias{predict_cells_time}
\title{Runtime calculation for PredictCells() function}
\usage{
predict_cells_time(
  dataset_name,
  data_dir,
  test_data_dir,
  test_data,
  model_dir,
  model_date,
  device,
  cluster_column,
  lr,
  margin,
  bs,
  epoch,
  knn_k,
  output_dim,
  dropout_prob,
  activation_function
)
}
\arguments{
\item{dataset_name}{Name of the training dataset}

\item{data_dir}{Path to reference data directory}

\item{test_data_dir}{Path to query/test data directory}

\item{test_data}{Name of the test dataset}

\item{model_dir}{Path to saved model directory}

\item{model_date}{Date when model was created (this helps CellFuse to load correct model)}

\item{device}{'cpu' or 'cuda'}

\item{cluster_column}{Column name containing cluster labels}

\item{lr}{Learning rate}

\item{margin}{Margin value for contrastive loss}

\item{bs}{Batch size}

\item{epoch}{Number of epochs}

\item{knn_k}{K for KNN classifier}

\item{output_dim}{Size of embedding output}

\item{dropout_prob}{Dropout probability}

\item{activation_function}{Activation function name}
}
\value{
Numeric value indicating runtime (in seconds).
}
\description{
Evaluates the computational runtime of the CellFuse query prediction step.
}
\details{
Useful for benchmarking CPU/GPU performance and scalability analysis.
}
