% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{IntegrateData}
\alias{IntegrateData}
\title{Integrate query cell types  with reference cell types}
\usage{
IntegrateData(ref_path, query_path, Celltype_col)
}
\arguments{
\item{ref_path}{Path to reference CSV file.}

\item{query_path}{Path to predicted query CSV file.}

\item{Celltype_col}{Character. Column containing cell-type labels.}
}
\value{
A data.frame containing normalized query marker values.
}
\description{
This function normalizes query data to reference data using PCA-space
quantile normalization, performed separately for each cell type.
}
\details{
Normalization is performed separately per cell type.
Only numeric marker columns are processed.
}
\examples{
\dontrun{
corrected <- IntegrateData(
  ref_path = "Reference_Data/CyTOF_train.csv",
  query_path = "Query_Data/CITEseq_test.csv",
  Celltype_col = "cluster.orig"
)
}


}
